async/await fixes
