+++ src/pages/characters/index.astro
@@ -2,15 +2,24 @@
 import Layout from "../../layouts/Layout.astro";
 import Content from "../../layouts/Content.astro";
 import CardCharacter from "../../components/CardCharacter.astro";
-import { getCharacters, getCharacterByIndex } from "../../services/json";
+import { getCharacters } from "../../services/database";
 
 const characters = await getCharacters();
 ---
 
 <Layout title="Lista de personajes">
     <Content>
-        <div class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 md:grid-cols-3 gap-5 w-full">
-            { characters.map((charPos) => <CardCharacter href={ `/characters/${charPos.id}` } char={ await getCharacterByIndex(charPos.pos) } /> ) }
+        <div
+            class="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 md:grid-cols-3 gap-5 w-full"
+        >
+            {
+                characters.map((item) => (
+                    <CardCharacter
+                        href={`/characters/${item.id}`}
+                        char={item}
+                    />
+                ))
+            }
         </div>
     </Content>
 </Layout>
